Don't mark outbound http spans as errors.
We log errors in all the cases that are legit (non-user) related errors
and so there's no need to also mark the span as errored. This prevents
user related issues (e.g., timeouts) from not being bubbled up to
telemetry as errors.

Additionally, we stop renaming the span after the http method which
isn't very helpful.

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>