enable connection pooling and reuse for outbound wasi-http
This updates `factor-outbound-http/wasi.rs` to use
`hyper_util::client::legacy::Client`, which does connection pooling by default.
`Client` requires a `tower_service::Service` implementation for creating new
connections, which in turn required moving code around to accomodate a different
flow of control.  It also forces us to work at a slightly higher level of
abstraction, so we don't have as much fine-grained control as before, notably:

- We can't make HTTP/2-specific request tweaks like we did before (e.g. removing
  or adding `host` headers, or tweaking the path-with-query).  I'm mainly
  leaving it up to `hyper-util` to do the right thing here :crossed_fingers:.

- We can't provide an `AbortOnDropJoinHandle` when constructing an
  `IncomingResponse` object, which may mean the connection continues to exist
  after the response is dropped, but that's kind of the point of connection
  pooling anyway.

Open questions: do we want to make this configurable, and if so, should it be
enabled or disabled by default?

Signed-off-by: Joel Dice <joel.dice@fermyon.com>