move the nginx HTTP2 bug workaround
We remove the `host` header from the request if present because (some versions
of?) nginx have a known bug such that they can't handle `host` headers in HTTP/2
requests.  However, the request interceptor, if present, may have its own
opinions about whether a `host` header should be set, and if it insists on
setting one we should honor that.  So I've moved the workaround to run _before_
the interceptor is run.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>