chore(examples): Update Spin Rust SDK and other dependencies
The Rust examples were all using Spin SDK 2.2.0, which targets a release candidate of WASIp2, not the final version. That means they only ran in Spin, not other WASI runtimes such as Wasmtime.

This commit updates all applicable examples to the latest Spin SDK and changes the build target to `wasm32-wasip2` to directly build components.

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>