add WASIp3 port of `wasi-http-streaming` test
This required adding the 0.3.0-rc-2025-08-15 WIT files to the tree so
`wit-bindgen` could use them.

I've also addressed Till's feedback to match only the exact 0.3.0 RC we support
when looking for the `wasi:http/handler` export.

Finally, this includes a temporary workaround for
https://github.com/bytecodealliance/wasmtime/issues/11703, which was uncovered
by the streaming test added in this commit.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>