chore(ci): Restrict use of rust-cache to longest-running jobs
Spin CI currently always exceeds the 10GB cache quota. That means that between creating a cache entry and trying to use it, it has pretty much always been LRU evicted, so we pay for creating the entries, and never benefit from them.

Or rather, that would be the situation if GitHub strictly enforced the cache limits. They apparently are planning on doing so starting in mid-October. Currently, we sometimes get eviction before any use, sometimes not.

This change makes it so that only the two longest-running jobs use the cache, and that it's only saved on `main`. In combination, we should get much better hit rates for those two jobs, which should hopefully mean that CI most of the time completes in a bit more than 30 minutes, instead of frequently taking close to an hour.

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>